Move useClusterStatus hook into separate file
