+++ src/components/MAPI/clusters/utils.ts
@@ -21,12 +21,6 @@ import * as capzv1beta1 from 'model/services/mapi/capzv1beta1';
 import * as corev1 from 'model/services/mapi/corev1';
 import * as infrav1alpha3 from 'model/services/mapi/infrastructurev1alpha3';
 import * as releasev1alpha1 from 'model/services/mapi/releasev1alpha1';
-import {
-  getIsImpersonatingNonAdmin,
-  getUserIsAdmin,
-} from 'model/stores/main/selectors';
-import { useMemo } from 'react';
-import { useSelector } from 'react-redux';
 import { mutate } from 'swr';
 import ErrorReporter from 'utils/errors/ErrorReporter';
 import { parseRFC822DateFormat } from 'utils/helpers';
@@ -859,55 +853,4 @@ export enum ClusterStatus {
   UpgradeInProgress = 'UPGRADE_IN_PROGRESS',
   UpgradeScheduled = 'UPGRADE_SCHEDULED',
   UpgradeAvailable = 'UPGRADE_AVAILABLE',
-}
-
-export function useClusterStatus(
-  cluster?: capiv1beta1.ICluster,
-  providerCluster?: ProviderCluster,
-  releases?: releasev1alpha1.IRelease[]
-): { status?: ClusterStatus; clusterUpdateSchedule?: IClusterUpdateSchedule } {
-  const isAdmin = useSelector(getUserIsAdmin);
-  const isImpersonatingNonAdmin = useSelector(getIsImpersonatingNonAdmin);
-  const provider = window.config.info.general.provider;
-  const { isConditionUnknown, isCreating, isUpgrading, isDeleting } =
-    getClusterConditions(cluster, providerCluster);
-
-  const isUpgradable = useMemo(
-    () =>
-      cluster &&
-      isClusterUpgradable(
-        cluster,
-        provider,
-        isAdmin && !isImpersonatingNonAdmin,
-        releases
-      ),
-    [cluster, provider, isAdmin, isImpersonatingNonAdmin, releases]
-  );
-
-  const clusterUpdateSchedule = getClusterUpdateSchedule(cluster);
-
-  if (typeof cluster === 'undefined') {
-    return { status: undefined };
-  }
-
-  switch (true) {
-    case isDeleting:
-      return { status: ClusterStatus.DeletionInProgress };
-
-    case isConditionUnknown:
-    case isCreating:
-      return { status: ClusterStatus.CreationInProgress };
-
-    case isUpgrading:
-      return { status: ClusterStatus.UpgradeInProgress };
-
-    case typeof clusterUpdateSchedule !== 'undefined':
-      return { status: ClusterStatus.UpgradeScheduled, clusterUpdateSchedule };
-
-    case isUpgradable:
-      return { status: ClusterStatus.UpgradeAvailable };
-
-    default:
-      return { status: ClusterStatus.Idle };
-  }
 }
